requirements.txt
asgiref==3.2.10
Django==3.1.2
django-autoslug==1.9.8
pytz==2020.1
sqlparse==0.4.1
